chore(cfs): narrow Ansible.verbosity from u64 to u8
CFS API caps Ansible verbosity at 0..=4 (-v through -vvvv); u64 was
unnecessarily wide and silently allowed unparseable values from the
wire to deserialize.

+++ src/types/cfs/session/mod.rs
@@ -30,7 +30,7 @@ pub struct Ansible {
   #[serde(skip_serializing_if = "Option::is_none")]
   pub limit: Option<String>,
   #[serde(skip_serializing_if = "Option::is_none")]
-  pub verbosity: Option<u64>,
+  pub verbosity: Option<u8>,
   #[serde(skip_serializing_if = "Option::is_none")]
   pub passthrough: Option<String>,
 }
